+++ mrmustard/math/lattice/strategies/vanilla.py
@@ -16,7 +16,7 @@ import numpy as np
 from numba import njit
 
 from mrmustard.math.lattice import paths, steps
-from mrmustard.typing import ComplexMatrix, ComplexTensor, ComplexVector
+from mrmustard.utils.typing import ComplexMatrix, ComplexTensor, ComplexVector
 
 SQRT = np.sqrt(np.arange(100000))